Add flag to retain v2 releases

+++ cmd/convert.go
@@ -32,6 +32,7 @@ var (
         tillerNamespace string
         label           string
 	dryRun          bool
+	keepv2Releases  bool
 )
 
 func newConvertCmd(out io.Writer) *cobra.Command {
@@ -52,6 +53,7 @@ func newConvertCmd(out io.Writer) *cobra.Command {
         flags.StringVarP(&tillerNamespace, "tiller-ns","t", "kube-system", "namespace of Tiller")
         flags.StringVarP(&label, "label", "l", "OWNER=TILLER", "label to select tiller resources by")
 	flags.BoolVar(&dryRun, "dry-run", false, "simulate a convert")
+	flags.BoolVar(&keepv2Releases, "keep-v2-releases", false, "v2 releases are retained after migration. By default, the v2 releases are removed")
 
 	return cmd
 
@@ -63,8 +65,9 @@ func run(cmd *cobra.Command, args []string) error {
 }
 
 // Convert coonverts helm 2 release into Hlem 3 relesae. It maps the Helm v2 release versions
-// of the releases into Helm v3 equivalent and stores the release versions. The underlying  Kubernetes resources
+// of the release into Helm v3 equivalent and stores the release versions. The underlying  Kubernetes resources
 // are untouched. Note: The namespaces of each release version need to exist in the Kubernetes  cluster.
+// The Helm 2 release is deleted by default, unless the '--keepv2Releases' flag is set.
 func Convert(releaseName string) error {
 	if dryRun {
 	        fmt.Printf("NOTE: This is in dry-run mode, the following actions will not be executed.\n")
@@ -93,17 +96,23 @@ func Convert(releaseName string) error {
                 versions = append(versions, version)
         }
         if !dryRun {
-                fmt.Printf("[Helm 3] Release \"%s\"  created.\n", releaseName)
+                fmt.Printf("[Helm 3] Release \"%s\" created.\n", releaseName)
 	}
 
-        fmt.Printf("[Helm 2] Release \"%s\" will be deleted.\n", releaseName)
-        if err := v2.DeleteReleaseVersions(releaseName, tillerNamespace, versions, dryRun); err != nil {
-                return err
-        }
-	if !dryRun {
-                fmt.Printf("[Helm 2] Release \"%s\"  deleted.\n", releaseName)
+	if !keepv2Releases {
+                fmt.Printf("[Helm 2] Release \"%s\" will be deleted.\n", releaseName)
+                if err := v2.DeleteReleaseVersions(releaseName, tillerNamespace, versions, dryRun); err != nil {
+                        return err
+                }
+	        if !dryRun {
+                        fmt.Printf("[Helm 2] Release \"%s\" deleted.\n", releaseName)
 
-                fmt.Printf("Release \"%s\" was converted successfully from Helm 2 to Helm 3.\n", releaseName)
+                        fmt.Printf("Release \"%s\" was converted successfully from Helm 2 to Helm 3.\n", releaseName)
+	        }
+	} else {
+		if !dryRun {
+			fmt.Printf("Release \"%s\" was converted successfully from Helm 2 to Helm 3. Note: the v2 releases still remain and should be removed to avoid conflicts with the migrated v3 releases.\n", releaseName)
+                }
 	}
 
         return nil
